requirements.txt
# Backend Server Requirements for Hand Sign Detection
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0
websockets==13.1

# Deep Learning
torch==2.5.1
torchvision==0.20.1
timm==1.0.11

# Image Processing
opencv-python==4.10.0.84
Pillow==11.1.0

# Utilities
numpy==2.2.3
python-multipart==0.0.20
